// src/api/model/apiPaginatedListApiPayment.ts
export interface ApiPaginatedListApiPayment { 
    /**
     * Response items.
     */
    items?: Array<ApiPayment>;
    /**
     * Count of all items satisfying 'paginatable' request.
     */
    count?: number;
    /**
     * Offset got from request
     */
    offset?: number;
    /**
     * Limit got from request
     */
    limit?: number;
}

/**
 * Namespace for property- and property-value-enumerations of ApiPaginatedListApiPayment.
 */
export namespace ApiPaginatedListApiPayment {
    /**
     * All properties of ApiPaginatedListApiPayment.
     */
    export enum Properties {
        /**
         * Response items.
         */
        items = 'items',
        /**
         * Count of all items satisfying 'paginatable' request.
         */
        count = 'count',
        /**
         * Offset got from request
         */
        offset = 'offset',
        /**
         * Limit got from request
         */
        limit = 'limit'
    }


    export function formMetadata() {
        return  {
            metadata: formMetadata,
            classname: 'ApiPaginatedListApiPayment',
            vars: [
                {
                    metadata: ApiPayment.formMetadata,
                    isReadOnly: false,
                    isEnum: false,
                    required: false,
                    name: 'items',
                    classname: 'ApiPaginatedListApiPayment',
                    dataType: 'Array&lt;ApiPayment&gt;',
                    isPrimitiveType: false,
                    isListContainer: true,
                    complexType: 'ApiPayment'
                },
                {
                    isReadOnly: false,
                    isEnum: false,
                    required: false,
                    name: 'count',
                    classname: 'ApiPaginatedListApiPayment',
                    dataType: 'number',
                    isPrimitiveType: true,
                    isListContainer: false,
                    complexType: ''
                },
                {
                    isReadOnly: false,
                    isEnum: false,
                    required: false,
                    name: 'offset',
                    classname: 'ApiPaginatedListApiPayment',
                    dataType: 'number',
                    isPrimitiveType: true,
                    isListContainer: false,
                    complexType: ''
                },
                {
                    isReadOnly: false,
                    isEnum: false,
                    required: false,
                    name: 'limit',
                    classname: 'ApiPaginatedListApiPayment',
                    dataType: 'number',
                    isPrimitiveType: true,
                    isListContainer: false,
                    complexType: ''
                },
            ],
            validators: {
                items: [
                ],
                count: [
                ],
                offset: [
                ],
                limit: [
                ],
            }
        }
    }

  // export const ApiPaginatedListApiPaymentValidationScheme = {
  //     validators: [],
  //     fields: {
  //               items: {
  //                   validators: []
  //               },
  //               count: {
  //                   validators: []
  //               },
  //               offset: {
  //                   validators: []
  //               },
  //               limit: {
  //                   validators: []
  //               },
  //     }
  // } as SimpleValidationScheme<ApiPaginatedListApiPayment>;


}